LA-611: Move upload + share toast message to HomeWidget
