+++ lib/presentation/widget/home/home_widget.dart
@@ -29,6 +29,8 @@
 //  3 and <http://www.linshare.org/licenses/LinShare-License_AfferoGPL-v3.pdf> for
 //  the Additional Terms applicable to LinShare software.
 
+import 'package:dartz/dartz.dart' as dartz;
+import 'package:domain/domain.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:flutter_svg/svg.dart';
@@ -38,6 +40,7 @@ import 'package:linshare_flutter_app/presentation/redux/actions/ui_action.dart';
 import 'package:linshare_flutter_app/presentation/redux/states/app_state.dart';
 import 'package:linshare_flutter_app/presentation/redux/states/ui_state.dart';
 import 'package:linshare_flutter_app/presentation/util/app_image_paths.dart';
+import 'package:linshare_flutter_app/presentation/util/app_toast.dart';
 import 'package:linshare_flutter_app/presentation/util/extensions/color_extension.dart';
 import 'package:linshare_flutter_app/presentation/util/router/route_paths.dart';
 import 'package:linshare_flutter_app/presentation/widget/myspace/my_space_widget.dart';
@@ -56,6 +59,7 @@ class _HomeWidgetState extends State<HomeWidget> {
   final _scaffoldKey = GlobalKey<ScaffoldState>();
   final homeViewModel = getIt<HomeViewModel>();
   final imagePath = getIt<AppImagePaths>();
+  final appToast = getIt<AppToast>();
 
   @override
   void dispose() {
@@ -83,10 +87,23 @@ class _HomeWidgetState extends State<HomeWidget> {
             onPressed: () => _scaffoldKey.currentState.openDrawer()),
       ),
       drawer: SideMenuDrawerWidget(),
-      body: StoreConnector<AppState, UIState>(
-        converter: (store) => store.state.uiState,
-        distinct: true,
-        builder: (context, uiState) => getHomeWidget(uiState)
+      body: Column(
+        children: [
+          handleUploadToastMessage(context),
+          handleShareDocumentToastMessage(context),
+          StoreConnector<AppState, Success>(
+              converter: (store) => store.state.uploadFileState.viewState.getOrElse(() => null),
+              distinct: true,
+              builder: (context, data) => handleUploadWidget(context, data)
+          ),
+          Expanded(
+            child: StoreConnector<AppState, UIState>(
+                converter: (store) => store.state.uiState,
+                distinct: true,
+                builder: (context, uiState) => getHomeWidget(uiState)
+            ),
+          ),
+        ],
       ),
     );
   }
@@ -119,6 +136,165 @@ class _HomeWidgetState extends State<HomeWidget> {
         );
       default:
         return getIt<MySpaceWidget>();
+    }
+  }
+
+  Widget handleUploadToastMessage(BuildContext context) {
+    return StoreConnector<AppState, dartz.Either<Failure, Success>>(
+        converter: (store) => store.state.uploadFileState.viewState,
+        distinct: true,
+        builder: (context, state) => state.fold((failure) {
+              if (failure is FilePickerFailure
+                  || failure is FileUploadFailure
+                  || failure is WorkGroupDocumentUploadFailure) {
+                appToast.showToast(AppLocalizations.of(context).upload_failure_text);
+              }
+              return SizedBox.shrink();
+            }, (success) {
+              if (success is FileUploadSuccess || success is WorkGroupDocumentUploadSuccess) {
+                appToast.showToast(AppLocalizations.of(context).upload_success_text);
+                homeViewModel.cleanUploadViewState();
+              }
+              return SizedBox.shrink();
+            }));
+  }
+
+  Widget handleShareDocumentToastMessage(BuildContext context) {
+    return StoreConnector<AppState, dartz.Either<Failure, Success>>(
+        converter: (store) => store.state.shareState.viewState,
+        distinct: true,
+        builder: (context, state) => state.fold((failure) {
+          if (failure is ShareDocumentFailure) {
+            appToast.showErrorToast(AppLocalizations.of(context).file_could_not_be_share);
+            homeViewModel.cleanShareViewState();
+          }
+          return SizedBox.shrink();
+        }, (success) {
+          if (success is ShareDocumentViewState) {
+            appToast.showToast(AppLocalizations.of(context).file_is_successfully_shared);
+            homeViewModel.cleanShareViewState();
+          } else if (success is ShareAfterUploadSuccess) {
+            appToast.showToast(_buildSharingMessage(context, success.recipients));
+            homeViewModel.cleanUploadViewState();
+            homeViewModel.cleanShareViewState();
+          }
+          return SizedBox.shrink();
+        }));
+  }
+
+  Widget handleUploadWidget(BuildContext context, [Success success]) {
+    if (success is UploadingProgress) {
+      return _buildUploadingFile(context, success.fileInfo.fileName, success.progress);
+    } else if (success is FilePickerSuccessViewState) {
+      return _buildPreparingUploadFile(context, success.fileInfo.fileName);
+    } else if (success is SharingAfterUploadState) {
+      return _buildSharingFileWidget(success);
+    } else {
+      return SizedBox.shrink();
+    }
+  }
+
+  Widget _buildUploadingFile(
+      BuildContext context, String fileName, int progress) {
+    return SizedBox(
+      key: Key('my_space_uploading_file'),
+      height: 58,
+      child: Container(
+        color: AppColor.mySpaceUploadBackground,
+        child: Column(children: [
+          _buildLinearProgress(context, progress),
+          Expanded(child: _buildUploadFileInfo(context, fileName, progress)),
+        ]),
+      ),
+    );
+  }
+
+  Widget _buildLinearProgress(BuildContext context, int progress) {
+    return SizedBox(
+      height: 4,
+      child: LinearProgressIndicator(
+        backgroundColor: AppColor.uploadProgressBackgroundColor,
+        valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
+        value: progress.toDouble() / 100,
+      ),
+    );
+  }
+
+  Widget _buildUploadFileInfo(
+      BuildContext context, String fileName, int progress) {
+    return Row(
+      children: [
+        Expanded(
+          child: Padding(
+            padding: EdgeInsets.only(left: 24),
+            child: Text(
+              'Uploading ' + fileName + ' ($progress)%...',
+              maxLines: 1,
+              style: TextStyle(fontSize: 14, color: Colors.white),
+            ),
+          ),
+        )
+      ],
+    );
+  }
+
+  Widget _buildPreparingUploadFile(BuildContext context, String fileName) {
+    return SizedBox(
+      key: Key('my_space_preparing_upload_file'),
+      height: 54,
+      child: Container(
+        color: AppColor.mySpaceUploadBackground,
+        child: Column(children: [
+          Expanded(
+              child: Row(
+                children: [
+                  Expanded(
+                    child: Padding(
+                      padding: EdgeInsets.only(left: 24),
+                      child: Text(
+                        AppLocalizations.of(context).upload_prepare_text,
+                        maxLines: 1,
+                        style: TextStyle(fontSize: 14, color: Colors.white),
+                      ),
+                    ),
+                  )
+                ],
+              )),
+        ]),
+      ),
+    );
+  }
+
+  Widget _buildSharingFileWidget(SharingAfterUploadState state) {
+    return SizedBox(
+      height: 58,
+      child: Container(
+        color: AppColor.mySpaceUploadBackground,
+        child: Align(
+          child: Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 24.0),
+            child: Text(
+              _buildSharingMessage(context, state.recipients),
+              maxLines: 1,
+              style: TextStyle(fontSize: 14, color: Colors.white),
+            ),
+          ),
+          alignment: Alignment.centerLeft,
+        ),
+      ),
+    );
+  }
+
+  String _buildSharingMessage(BuildContext context, List<AutoCompleteResult> recipients) {
+    final shareSinglePerson = recipients.length == 1;
+    if (shareSinglePerson) {
+      return AppLocalizations.of(context).sharing_single_after_uploaded_success(
+          recipients.first.getSuggestionDisplayName()
+      );
+    } else {
+      return AppLocalizations.of(context).sharing_multiple_after_uploaded_success(
+          recipients.length
+      );
     }
   }
 }
